+++ src/radiofuncs.h
@@ -16,5 +16,7 @@
 
 STRPTR getRadioStations(struct filters, int);
 STRPTR getRadioTrendStations(void);
+STRPTR getRadioPopularStations(void);
 void getRadioList(STRPTR, int);
 void getRadioTrendList(STRPTR, int);
+void getRadioPopularList(STRPTR, int);